refactor: move primitives to separate file
